+++ CHANGELOG.md
@@ -1,4 +1,5 @@
  * Updated FFmpeg from 3.4.2 to 4.0.1
+ * Added API methods to get PES information as a JSON string
  * Separated integration tests using Maven Failsafe Plugin
  * De-clutter build output making warnings easier to spot